requirements.txt
fastapi
uvicorn
redis
python-dotenv
langgraph
langchain-core
langchain-ollama
ollama
